backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
Pillow==9.4.0
pytesseract==0.3.10
gunicorn==20.1.0
python-dotenv==1.0.0 